pyproject.toml
[project]
name = "multi-step-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
]
